### Resolves

Faster loading.

### Proposed Changes

Check if toolbox needs to be updated with new XML during updateToolbox.

### Reason for Changes

src/containers/Blocks.jsx:updateToolbox may be called when the xml hasn't changed. Calling this.ScratchBlocks.updateToolbox with the unchanged XML does a lot of work for no change.

### Benchmark Data

|  device | before (ms) | after (ms) | difference (ms) | change (%) |
| --- | --- | --- | --- | --- |
|  chrome | 117 | 1 | 116 | 99.15% |
|  firefox | 143 | 1 | 142 | 99.30% |
|  safari | 99 | 1 | 98 | 98.79% |
|  chromebook | 425 | 2 | 422 | 99.43% |