@paulkaplan 2019-03-22
The integration test that is failing is relevant here. I believe the toolbox not being updated is causing the variables not to appear.