@mzgoddard 2019-03-25
The second failing test is caused by the first failing test. On its own it doesn't fail.

The first is failing I think because the toolbox rendering is not deterministically built from the string passed to updateToolbox. The data category determines its variable list from the variables in the main workspace's variable map instead of the variables in the xml passed to it. The failing tests is failing at the last assertion. The first check for the variable passes as its populated from a updateToolbox call after the default project loads and that includes the default workspace. The second check fails because the toolbox rendering at inject time does not correctly render the toolbox. It is missing the variables that have not been loaded into the workspace. A updateToolbox fires later after the workspace is updated but the correct toolbox xml has already been rendered. I think there is a fix here we can do but I'll have to figure it out tomorrow.