@mzgoddard 2019-03-26
I added some work that tracks when the workspace has first rendered. Until that first render we can't authoritatively or deterministically render the toolbox. If we render only toolbox changes after the workspace has rendered it'll be all up to date.

-----

Given the increased work for the toolbox render after the screen transition I wonder if we might switch the timeout in blocks to promises or another same-microtask-queue behaviour. That'll move the render to before the new screen is displayed and there won't be the larger flash caused by this change. (There is currently a flash. Just I supposed less noticeable.)

I could add something like ...

```diff
diff --git a/src/containers/blocks.jsx b/src/containers/blocks.jsx
index c366da45..f9cc3ba1 100644
--- a/src/containers/blocks.jsx
+++ b/src/containers/blocks.jsx
@@ -180,13 +180,19 @@ class Blocks extends React.Component {
     componentWillUnmount () {
         this.detachVM();
         this.workspace.dispose();
-        clearTimeout(this.toolboxUpdateTimeout);
+        this.toolboxUpdateTimeout && this.toolboxUpdateTimeout.cancel();
     }
     requestToolboxUpdate () {
-        clearTimeout(this.toolboxUpdateTimeout);
-        this.toolboxUpdateTimeout = setTimeout(() => {
-            this.updateToolbox();
-        }, 0);
+        this.toolboxUpdateTimeout && this.toolboxUpdateTimeout.cancel();
+        let running = true;
+        this.toolboxUpdateTimeout = Promise.resolve().then(() => {
+            if (running) {
+                this.updateToolbox();
+            }
+        });
+        this.toolboxUpdateTimeout.cancel = () => {
+            running = false;
+        };
     }
     setLocale () {
         this.ScratchBlocks.ScratchMsgs.setLocale(this.props.locale);
```