@paulkaplan 2019-04-18
@mzgoddard I think that the large flash from empty -> full of blocks doesn't seem like an acceptable change in user experience. The alternative you suggested seems worth a try.

Also I don't know what the benchmark numbers are referring to. Could you give a specific start/end that you are measuring from so I can replicate?