@mzgoddard 2019-04-26
The benchmark numbers refer to the first toolbox render when blocks is injected into the dom in the componentDidMount handle. I measured with a `start = Date.now(); /* inject blocks with uninitialized toolbox xml */ console.log('time', Date.now() - start);` kind of thing.