@mzgoddard 2019-04-26
I thought I check for comments on Monday and should have seen your response @paulkaplan. Sorry for the delay. I'll have the mentioned patch in, in a bit.